extract log schema to module
